change the mark argument's type to be u32

+++ src/listeners/tcp.rs
@@ -64,7 +64,7 @@ async fn tcp_handle_connection(
     log::info!("source addr: {src_addr}");
     log::info!("target addr: {target_addr}");
 
-    let mut upstream_conn = create_upstream_conn(src_addr, target_addr, args.mark as u32).await?;
+    let mut upstream_conn = create_upstream_conn(src_addr, target_addr, args.mark).await?;
     log::info!("created the upstream connection");
 
     conn.set_nodelay(true)?;
